heavy cleanup of clap and configuration structures
